// src/api-client/model/rentalStatusWorkerLogDetailRentalStatus.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * 빌릴게(Billilge) API
 * 국민대학교 소프트웨어융합대학 복지물품 대여 서비스
 * OpenAPI spec version: v1
 */

/**
 * 대여 상태
 */
export type RentalStatusWorkerLogDetailRentalStatus = typeof RentalStatusWorkerLogDetailRentalStatus[keyof typeof RentalStatusWorkerLogDetailRentalStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RentalStatusWorkerLogDetailRentalStatus = {
  PENDING: 'PENDING',
  CANCEL: 'CANCEL',
  CONFIRMED: 'CONFIRMED',
  REJECTED: 'REJECTED',
  RENTAL: 'RENTAL',
  RETURN_PENDING: 'RETURN_PENDING',
  RETURN_CONFIRMED: 'RETURN_CONFIRMED',
  RETURNED: 'RETURNED',
} as const;